Move Direction class to separate file
